chore(docs): move experiemnet to experiments dir and add readme
