+++ experiments/ao-connect/README.md
@@ -1,0 +1,3 @@
+## spawn
+
+Uses `ao-connect` to spawn a new process with a prescribed scheduler.
